pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "japonica-crm"
version = "0.1.0"
description = "Japonica CRM monolith scaffold"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "celery>=5.4.0,<6.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "orjson>=3.10.0,<4.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "python-multipart>=0.0.12,<1.0.0",
  "redis>=5.2.0,<6.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0"
]

[dependency-groups]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "ruff>=0.8.0,<1.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

